Handle contact form submit by opening new browser window using `mailto` protocol
Right now the "mailto" protocol seems to be a convenient method of sending
email from the client side to the developer but this includes opening of a
another window with Email form, which in our case is GMail. We would like to
have a direct mail submittion without any physical contact between a service
and the client.

Signed-off-by: Ayush Joshi <ayush854032@gmail.com>